took request stuff out of WebServerProg
