+++ Makefile
@@ -5,7 +5,7 @@ NAME = WebServ
 CC = c++
 
 # compiler flags
-CFLAGS =  -std=c++98 -Wall -Werror -Wextra -MMD
+CFLAGS =  -std=c++17 -Wall -Werror -Wextra -MMD
 DEBUG_FLAGS = -g
 
 # src files and path for them
@@ -14,7 +14,7 @@ INCS := $(wildcard $(addsuffix /*.hpp, $(VPATH)))
 INCLUDE = $(addprefix -I, $(VPATH))
 
 # src files
-SRC := main.cpp WebServerProg.cpp api_helpers.cpp api_get.cpp parser.cpp validateServers.cpp
+SRC := main.cpp WebServerProg.cpp api_helpers.cpp api_get.cpp parser.cpp validateServers.cpp  request.cpp
 
 # obj files and path for them
 OBJ_DIR = ./obj
