+++ src/WebServ/WebServerProg.hpp
@@ -1,5 +1,3 @@
-
-
 #ifndef WEBSERVERPROG_H
 #define WEBSERVERPROG_H
